rename folder and update main.go
