+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/MGTheTrain/terraform-provider-blob-services/layers"
+	layers "github.com/MGTheTrain/terraform-provider-blob-services/mgtt"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
